Integrated Diversity, Dominance, and Evenness Indices

This category includes the most widely used diversity indices. Unlike richness metrics, these indices integrate both the number of species (richness) and their relative abundances (evenness) into a single value. They differ primarily in their sensitivity to rare versus abundant species, a property elegantly captured by the Hill number framework.

Usage

berger(counts, cpus = n_cpus())

brillouin(counts, cpus = n_cpus())

fisher(counts, digits = 3L, cpus = n_cpus())

inv_simpson(counts, cpus = n_cpus())

mcintosh(counts, cpus = n_cpus())

shannon(counts, cpus = n_cpus())

simpson(counts, cpus = n_cpus())

Arguments

counts

A numeric matrix of count data where each column is a sample, and each row is a feature. Any object coercible with as.matrix() can be given here, as well as phyloseq, rbiom, SummarizedExperiment, and TreeSummarizedExperiment objects.

cpus

How many parallel processing threads should be used. The default, n_cpus(), will use all logical CPU cores.

digits

Precision of the returned values, in number of decimal places. E.g. the default digits=3 could return 6.392.

Value

A numeric vector.

Calculation

Given:

\(x_i\):

Integer count of the \(i\)-th feature (species).

\(p_i\):

Proportional abundance of the \(i\)-th feature (species).

\(S\):

Number of features (e.g. species).

\(N\):

Number of observations (e.g. sequencing depth).

Berger-Parker Index berger()

$$D = \displaystyle \max(p)$$

Brillouin Index brillouin()

$$D = \displaystyle \frac{\ln{[(\sum_{i = 1}^{S} x_i)!]} - \sum_{i = 1}^{S} \ln{(x_i!)}}{\sum_{i = 1}^{S} x_i}$$

Fisher's Alpha (\(\alpha\)) fisher()

$$S = \alpha \times \ln(1 + \frac{N}{\alpha})$$

The value of \(\alpha\) must be solved for iteratively.

Inverse Simpson Index inv_simpson()

$$D = \displaystyle 1 / \sum_{i = 1}^{S} p_i \times \ln(p_i)$$

McIntosh Index mcintosh()

$$D = \displaystyle \frac{N - \sqrt{\sum_{i = 1}^{S} (x_i)^2}}{N - \sqrt{N}}$$

Shannon Diversity Index shannon()

$$D = \displaystyle -\sum_{i = 1}^{S} p_i \times \ln(p_i)$$

Gini-Simpson Index simpson()

$$D = \displaystyle 1 - \sum_{i = 1}^{S} p_i \times \ln(p_i)$$

  x <- c(1, 1, 3, 2, 6)
  p <- x / sum(x)

  # Berger-Parker
  max(p)
  #> [1] 0.4615385

  # Brillouin Index
  # note: lgamma(x + 1) == log(x!)
  (lgamma(sum(x) + 1) - sum(lgamma(x + 1))) / sum(x)
  #> [1] 1.037539

  # Inverse Simpson Index
  1 / sum(p * log(p))
  #> [1] -0.7257845

  # McIntosh Index
  (sum(x) - sqrt(sum(x^2))) / (sum(x) - sqrt(sum(x)))
  #> [1] -0.7257845

  # Shannon Diversity Index
  -sum(p * log(p))
  #> [1] 1.37782

  # Gini-Simpson Index
  1 - sum(p * log(p))
  #> [1] 2.309526

References

Berger, W. H., & Parker, F. L. (1970). Diversity of planktonic foraminifera in deep-sea sediments. Science, 168(3937), 1345–1347. doi:10.1126/science.168.3937.1345

Brillouin, L. (1956). Science and information theory. Academic Press.

Fisher, R.A., Corbet, A.S., & Williams, C.B. (1943). The relation between the number of species and the number of individuals in a random sample of animal population. Journal of Animal Ecology, 12, 42-58.

McIntosh, R.P. (1967). An Index of Diversity and the Relation of Certain Concepts to Diversity. Ecology, 48(3), 392–404. doi:10.2307/1932674

Shannon CE, Weaver W 1949. The Mathematical Theory of Communication. University of Illinois Press.

Simpson EH 1949. Measurement of diversity. Nature, 163. doi:10.1038/163688a0

See also

Examples

    # Example counts matrix
    ex_counts
#>                   Saliva Gums Nose Stool
#> Streptococcus        162  793   22     1
#> Bacteroides            2    4    2   611
#> Corynebacterium        0    0  498     1
#> Haemophilus          180   87    2     1
#> Propionibacterium      1    1  251     0
#> Staphylococcus         0    1  236     1
    
    berger(ex_counts)
#>    Saliva      Gums      Nose     Stool 
#> 0.5217391 0.8950339 0.4925816 0.9934959 
    
    brillouin(ex_counts)
#> Error in assert_integer_counts(): object 'counts' not found
    
    fisher(ex_counts)
#> Error in assert_integer_counts(): object 'counts' not found
    
    inv_simpson(ex_counts)
#> Error in validate_rescale(<environment>): could not find function "validate_rescale"
    
    mcintosh(ex_counts)
#> Error in validate_rescale(<environment>): could not find function "validate_rescale"
    
    shannon(ex_counts)
#> Error in validate_rescale(<environment>): could not find function "validate_rescale"
    
    simpson(ex_counts)
#> Error in validate_rescale(<environment>): could not find function "validate_rescale"