Information-Theoretic and Probabilistic Distances

This category of metrics is derived from information theory and probability theory. They conceptualize sample composition vectors as discrete probability distributions and measure the dissimilarity by quantifying the difference between these distributions.

Usage

divergence(counts, pairs = NULL, cpus = n_cpus())

jensen(counts, pairs = NULL, cpus = n_cpus())

jsd(counts, pairs = NULL, cpus = n_cpus())

prob_symm(counts, pairs = NULL, cpus = n_cpus())

squared_chisq(counts, pairs = NULL, cpus = n_cpus())

topsoe(counts, pairs = NULL, cpus = n_cpus())

Arguments

counts

A numeric matrix of count data where each column is a sample, and each row is a feature. Any object coercible with as.matrix() can be given here, as well as phyloseq, rbiom, SummarizedExperiment, and TreeSummarizedExperiment objects.

pairs

Which combinations of samples should distances be calculated for? The default value (NULL) calculates all-vs-all. Provide a numeric or logical vector specifying positions in the distance matrix to calculate. See examples.

cpus

How many parallel processing threads should be used. The default, n_cpus(), will use all logical CPU cores.

Value

A dist object.

Details

Divergence-Based Metrics: Jensen-Shannon and Topsoe

The Jensen-Shannon Divergence (JSD) is a widely used method for comparing probability distributions. It is a symmetric and bounded (ranging from 0 to 1) version of the more famous but asymmetric Kullback-Leibler (KL) divergence. JSD effectively measures the "information gain" or "surprise" in discriminating between the two distributions. Its square root, the Jensen-Shannon distance, is a true metric.

The Topsoe distance is another metric from the same family of information-theoretic measures, sharing similar properties and applications. These metrics are robust for proportional data and are less sensitive to outliers than some geometric distances.

Chi-Squared Based Distances

The Squared Chi-Squared and Probabilistic Symmetric Chi-Squared distances are based on the chi-squared (\(\chi^2\)) statistic, which is commonly used in contingency table tests to assess the independence of two categorical variables. In the context of beta diversity, it compares the observed species proportions in two communities against what would be expected if they were drawn from the same underlying distribution. This makes it particularly sensitive to differences in the relative proportions of species, even when absolute abundances are low. It is a powerful tool for detecting associations and differences in frequency data.

Calculation

Given:

\(x\):

Feature counts for the first sample.

\(y\):

Feature counts for the second sample.

\(n\):

The number of features.

Jensen-Shannon divergence (JSD) jsd()

$$D = \displaystyle \frac{1}{2}\left[\sum_{i=1}^{n}x_i\ln\left(\frac{2x_i}{x_i + y_i}\right) + \sum_{i}y_i\ln\left(\frac{2y_i}{x_i + y_i}\right)\right]$$

Jensen-Shannon distance jensen()

$$D = \displaystyle \sqrt{\frac{1}{2}\left[\sum_{i=1}^{n}x_i\ln\left(\frac{2x_i}{x_i + y_i}\right) + \sum_{i}y_i\ln\left(\frac{2y_i}{x_i + y_i}\right)\right]}$$

Probabilistic Symmetric Chi-Squared distance prob_symm()

$$D = \displaystyle 2\sum_{i=1}^{n}\frac{(x_i - y_i)^2}{x_i + y_i}$$

Squared Chi-Squared distance squared_chisq()

$$D = \displaystyle \sum_{i=1}^{n}\frac{(x_i - y_i)^2}{x_i + y_i}$$

Topsoe distance topsoe()

$$D = \displaystyle \left[\sum_{i=1}^{n}x_i\ln\left(\frac{2x_i}{x_i + y_i}\right) + \sum_{i}y_i\ln\left(\frac{2y_i}{x_i + y_i}\right)\right]$$

  x <- c(4, 1, 3, 2, 6)
  y <- c(1, 8, 1, 1, 5)

  # Jensen-Shannon Divergence (JSD)
  sum(x * log(2 * x / (x+y)), y * log(2 * y / (x+y))) / 2
  #> [1] 2.400612

  # Jensen-Shannon distance (jensen)
  sqrt(sum(x * log(2 * x / (x+y)), y * log(2 * y / (x+y))) / 2)
  #> [1] 2.400612

  # Topsoe
  sum(x * log(2 * x / (x+y)), y * log(2 * y / (x+y)))
  #> [1] 4.801224

  # Divergence
  2 * sum((x - y)^2 / (x + y)^2)
  #> [1] 2.668628

  # Probabilistic Symmetric Chi-Squared
  2 * sum((x - y)^2 / (x + y))
  #> [1] 17.33737

  # Squared Chi-Squared
  sum((x - y)^2 / (x + y))
  #> [1] 8.668687

References

Cha, S.-H. (2007). Comprehensive survey on distance/similarity measures between probability density functions. International Journal of Mathematical Models and Methods in Applied Sciences, 1(4), 300–307.

Levy, A., Shalom, B. R., & Chalamish, M. (2024). A guide to similarity measures. arXiv. doi:10.48550/arXiv.2408.07706v1

Examples

    # Example counts matrix
    ex_counts
#>                   Saliva Gums Nose Stool
#> Streptococcus        162  793   22     1
#> Bacteroides            2    4    2   611
#> Corynebacterium        0    0  498     1
#> Haemophilus          180   87    2     1
#> Propionibacterium      1    1  251     0
#> Staphylococcus         0    1  236     1
    
    divergence(ex_counts)
#> Error in validate_normalized(<environment>): could not find function "validate_normalized"
    
    jensen(ex_counts)
#> Error in validate_normalized(<environment>): could not find function "validate_normalized"
    
    jsd(ex_counts)
#> Error in validate_normalized(<environment>): could not find function "validate_normalized"
    
    prob_symm(ex_counts)
#> Error in validate_normalized(<environment>): could not find function "validate_normalized"
    
    squared_chi(ex_counts)
#> Error in squared_chi(ex_counts): could not find function "squared_chi"
    
    topsoe(ex_counts)
#> Error in validate_normalized(<environment>): could not find function "validate_normalized"
    
    # Only calculate distances for A vs all.
    divergence(ex_counts, pairs = 1:3)
#> Error in validate_normalized(<environment>): could not find function "validate_normalized"