Horn-Morisita dissimilarity

A similarity index based on Simpson's diversity index, suitable for abundance data.

Usage

horn(counts, norm = "none", margin = 1L, pairs = NULL, cpus = n_cpus())

Arguments

counts

A numeric matrix of count data (samples \(\times\) features). Typically contains absolute abundances (integer counts), though proportions are also accepted.

norm

Normalize the incoming counts. Options are:

'none': No transformation.

'percent': Relative abundance (sample abundances sum to 1).

'binary': Unweighted presence/absence (each count is either 0 or 1).

'clr': Centered log ratio.

Default: 'none'.

margin

The margin containing samples. 1 if samples are rows, 2 if samples are columns. Ignored when counts is a special object class (e.g. phyloseq). Default: 1

pairs

Which combinations of samples should distances be calculated for? The default value (NULL) calculates all-vs-all. Provide a numeric or logical vector specifying positions in the distance matrix to calculate. See examples.

cpus

How many parallel processing threads should be used. The default, n_cpus(), will use all logical CPU cores.

Details

The Horn-Morisita dissimilarity is defined as: $$1 - \frac{2\sum_{i=1}^{n}P_{i}Q_{i}}{\sum_{i=1}^{n}P_i^2 + \sum_{i=1}^{n}Q_i^2}$$

Where:

\(P_i\), \(Q_i\) : Proportional abundances of the \(i\)-th feature.

\(n\) : The number of features.

Base R Equivalent:

z <- sum(x^2) / sum(x)^2 + sum(y^2) / sum(y)^2
1 - ((2 * sum(x * y)) / (z * sum(x) * sum(y)))

Input Types

The counts parameter is designed to accept a simple numeric matrix, but seamlessly supports objects from the following biological data packages:

phyloseq

rbiom

SummarizedExperiment

TreeSummarizedExperiment

For large datasets, standard matrix operations may be slow. See vignette('performance') for details on using optimized formats (e.g. sparse matrices) and parallel processing.

References

Horn, H. S. (1966). Measurement of "overlap" in comparative ecological studies. The American Naturalist, 100(914), 419-424. doi:10.1086/282436

See also

Examples

    horn(ex_counts)
#>          Saliva      Gums      Nose
#> Gums  0.2765026                    
#> Nose  0.9718747 0.9654300          
#> Stool 0.9900374 0.9932115 0.9952692